# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deposit-address:
    get:
      operationId: getDepositAddress
      tags: [wallet]
      summary: Get or generate a deposit address for a coin+network
      security: [{ cookieAuth: [] }]
      parameters:
        - name: coinId
          in: query
          required: true
          schema: { type: integer }
        - name: networkId
          in: query
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DepositAddress" }
        "404": { description: Network not found }
...
  /networks:
    get:
      operationId: listNetworks
      tags: [wallet]
      summary: List deposit/withdrawal networks (optionally filtered by coinId)
      parameters:
        - name: coinId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Network" }

  # ─── Earn ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cx_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DepositAddress:
      type: object
      properties:
        address: { type: string }
        memo: { type: string, nullable: true }
        networkId: { type: integer }
        coinId: { type: integer }
        status: { type: string }
      required: [address, networkId, coinId, status]
    Network:
      type: object
      properties:
        id: { type: integer }
        coinId: { type: integer }
        name: { type: string }
        chain: { type: string }
        contractAddress: { type: string, nullable: true }
        minDeposit: { type: string }
        minWithdrawal: { type: string }
        withdrawalFee: { type: string }
        status: { type: string, enum: [active, maintenance, inactive] }
        confirmations: { type: integer }
      required: [id, coinId, name, chain, status]
...